refactor: reorganize user routes and controllers for improved modularity and clarity, removing deprecated user management file
